requirements.txt
lmdb
matplotlib
torch
torchvision
scikit_image
numpy
scipy
chars2vec
natsort
fire
nltk
six
colour
opencv_contrib_python
Pillow
Polygon3
PyYAML
scikit-image
onnx
onnxruntime-gpu
